refactor: extract findFeatureDir to shared utility
Extract duplicated findFeatureDir function from 5 files to a shared
utility module, eliminating 107 lines of code duplication.

Files refactored:
- src/scheduler/specflow/gates/quality-gate.ts
- src/scheduler/specflow/orchestrator.ts
- src/scheduler/specflow/phases/complete.ts
- src/scheduler/specflow/phases/implement.ts
- src/scheduler/specflow-runner.ts
- src/commands/specflow-queue.ts

Created shared utility:
- src/scheduler/specflow/utils/find-feature-dir.ts

All instances now use the same implementation with consistent behavior:
case-insensitive prefix matching on feature directories.

Tests: 38 SpecFlow tests passing
